+++ components/Project/BookList/BookProperties/BookProperties.js
@@ -1,21 +1,21 @@
-import { useEffect, useState } from 'react'
-import Link from 'next/link'
+import { useEffect, useMemo, useState } from 'react'
 import axios from 'axios'
 import toast, { Toaster } from 'react-hot-toast'
-import ReactTextareaAutosize from 'react-textarea-autosize'
 
 import Modal from 'components/Modal'
+import Property from './Property'
 
-function PropertiesOfBook({
-  projectId,
-  book,
+function BookProperties({
+  bookCode,
   openDownloading,
   setOpenDownloading,
   t,
   type,
   user,
   mutateBooks,
+  books,
 }) {
+  const book = useMemo(() => books?.find((el) => el.code === bookCode), [bookCode, books])
   const [properties, setProperties] = useState()
   useEffect(() => {
     setProperties(book?.properties)
@@ -84,40 +84,4 @@ function PropertiesOfBook({
     </Modal>
   )
 }
-export default PropertiesOfBook
-
-function Property({ t, property, content, type, updateProperty }) {
-  const [propertyContent, setPropertyContent] = useState()
-  useEffect(() => {
-    setPropertyContent(content)
-  }, [content])
-  const additionalLinks = {
-    intro: 'https://git.door43.org/ru_gl/ru_obs/raw/branch/master/content/front/intro.md',
-    back: 'https://git.door43.org/ru_gl/ru_obs/raw/branch/master/content/back/intro.md',
-  }
-  return (
-    <>
-      <div className="inline-block mr-2">
-        {t(`book-properties:${property}${type === 'obs' ? '_obs' : ''}`)}
-      </div>
-      {additionalLinks[property] && (
-        <Link href={additionalLinks[property]}>
-          <a title={additionalLinks[property]} target="_blank" className="text-blue-450">
-            ?
-          </a>
-        </Link>
-      )}
-
-      <ReactTextareaAutosize
-        maxRows="5"
-        className="input"
-        placeholder={t(
-          `book-properties:${property}_placeholder${type === 'obs' ? '_obs' : ''}`
-        )}
-        value={propertyContent}
-        onChange={(e) => setPropertyContent(e.target.value)}
-        onBlur={() => updateProperty(propertyContent, property)}
-      />
-    </>
-  )
-}
+export default BookProperties
