+++ components/Project/index.js
@@ -1,5 +1,5 @@
 export { default as Project } from './Project'
 export { default as BookCreate } from './BookCreate'
 export { default as ChapterList } from './ChapterList'
-export { default as PropertiesOfBook } from './PropertiesOfBook'
-export { default as Download } from './Download'
+export { default as PropertiesOfBook } from './BookList/BookProperties/BookProperties'
+export { default as Download } from './BookList/Download'
